refactor(tests): consolidate `SRXManagerTest` into `SRXTest`
- Removed `SRXManagerTest` and integrated its test cases into `SRXTest`.
- Ensured consistent structure by organizing `SRXManager` unit tests under a nested class in `SRXTest`.
