+++ test/src/org/omegat/core/segmentation/SRXTest.java
@@ -4,7 +4,7 @@
           glossaries, and translation leveraging into updated projects.
 
  Copyright (C) 2016 Aaron Madlon-Kay
-               2024 Hiroshi Miura
+               2024-2026 Hiroshi Miura
                2025 Thomas Cordonnier
                Home page: https://www.omegat.org/
                Support center: https://omegat.org/support
@@ -43,6 +43,7 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Objects;
 
+import org.junit.Assert;
 import org.junit.Test;
 import org.junit.experimental.runners.Enclosed;
 import org.junit.rules.TemporaryFolder;
@@ -239,12 +240,76 @@ public final class SRXTest {
             Files.writeString(segmentConf, xmlContent);
             SRX srx = SRXManager.loadFromDir(segmentConf.getParent().toFile());
             assertNotNull(srx);
-            assertFalse(new File(tmpDir, "test-file").exists()); // true would
-                                                                 // mean that
-                                                                 // the
-                                                                 // vulnerability
-                                                                 // is still
-                                                                 // here!
+            assertFalse(new File(tmpDir, "test-file").exists());
+            // true would mean that the vulnerability is still here!
+        }
+    }
+
+    /**
+     * Unit tests for the {@link SRXManager#getDefault()} method. This method loads
+     * the default segmentation rules and configures the SRX object accordingly.
+     */
+    public static class SRXManagerUnitTest {
+
+        /**
+         * Test that the {@link SRXManager#getDefault()} method successfully loads
+         * the default SRX object.
+         */
+        @Test
+        public void testGetDefaultLoadsSRXSuccessfully() throws IOException {
+            SRX srx = SRXManager.getDefault();
+            Assert.assertNotNull("The SRX object should not be null.", srx);
+        }
+
+        /**
+         * Test that the {@link SRXManager#getDefault()} method sets
+         * includeEndingTags to true.
+         */
+        @Test
+        public void testGetDefaultIncludeEndingTagsIsTrue() throws IOException {
+            SRX srx = SRXManager.getDefault();
+            Assert.assertTrue("The includeEndingTags property should be true.", srx.isIncludeEndingTags());
+        }
+
+        /**
+         * Test that the {@link SRXManager#getDefault()} method sets segmentSubflows
+         * to true.
+         */
+        @Test
+        public void testGetDefaultSegmentSubflowsIsTrue() throws IOException {
+            SRX srx = SRXManager.getDefault();
+            Assert.assertTrue("The segmentSubflows property should be true.", srx.isSegmentSubflows());
+        }
+
+        /**
+         * Test that the {@link SRXManager#getDefault()} method returns an SRX
+         * object with a non-null and non-empty version string.
+         */
+        @Test
+        public void testGetDefaultVersion() throws IOException {
+            SRX srx = SRXManager.getDefault();
+            Assert.assertEquals("The SRX version should be 2.0.", "2.0", srx.getVersion());
+        }
+
+        /**
+         * Test that the {@link SRXManager#getDefault()} method returns an SRX
+         * object with non-null mapping rules.
+         */
+        @Test
+        public void testGetDefaultMappingRulesIsNotNull() throws IOException {
+            SRX srx = SRXManager.getDefault();
+            Assert.assertNotNull("The mapping rules should not be null.", srx.getMappingRules());
+        }
+
+        /**
+         * Test that the {@link SRXManager#getDefault()} method returns an SRX
+         * object with mapping rules containing at least one rule.
+         */
+        @Test
+        public void testGetDefaultMappingRulesHas18() throws IOException {
+            SRX srx = SRXManager.getDefault();
+            Assert.assertEquals("The mapping rules should contain at least one rule.", 18,
+                    srx.getMappingRules().size());
         }
     }
